pyproject.toml
[project]
name = "Backend_login_django"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "LucasGabrielAntonete", email = "lucasantonete@gmail.com"},
]
dependencies = [
    "django>=5.0.3",
    "setuptools>=69.2.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.3.1",
]
requires-python = "==3.11.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
